OD19DEVAT4-37 OD19DEVAT4-36 1) Update AttrtypeProperties and ObjtypeProperties from byte type of constants to BigInteger 2) Update CategoryDAO and GradeCategoryDAO according to changes in AttrtypeProperties and ObjtypeProperties

+++ src/main/java/com/netcracker/testerritto/properties/ObjtypeProperties.java
@@ -1,15 +1,17 @@
 package com.netcracker.testerritto.properties;
 
+import java.math.BigInteger;
+
 public interface ObjtypeProperties {
-  byte USER = 1;
-  byte GROUP = 2;
-  byte REMARK = 3;
-  byte TEST = 4;
-  byte RESULT = 5;
-  byte REPLY = 6;
-  byte IMAGE = 7;
-  byte CATEGORY = 8;
-  byte GRADE_CATEGORY = 9;
-  byte QUESTION = 10;
-  byte ANSWER = 11;
+  BigInteger USER = BigInteger.valueOf(1);
+  BigInteger GROUP = BigInteger.valueOf(2);
+  BigInteger REMARK = BigInteger.valueOf(3);
+  BigInteger TEST = BigInteger.valueOf(4);
+  BigInteger RESULT = BigInteger.valueOf(5);
+  BigInteger REPLY = BigInteger.valueOf(6);
+  BigInteger IMAGE = BigInteger.valueOf(7);
+  BigInteger CATEGORY = BigInteger.valueOf(8);
+  BigInteger GRADE_CATEGORY = BigInteger.valueOf(9);
+  BigInteger QUESTION = BigInteger.valueOf(10);
+  BigInteger ANSWER = BigInteger.valueOf(11);
 }
